backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0

# LangChain / LangGraph
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-experimental>=0.3.0
langgraph>=0.2.0

# LlamaIndex
llama-index-core>=0.12.0
llama-index-embeddings-openai>=0.3.0
llama-index-retrievers-bm25>=0.5.0
rank-bm25>=0.2.2

# Utilities
tiktoken>=0.7.0
html2text>=2020.1.16
httpx>=0.27.0
pydantic>=2.0.0
PyYAML>=6.0
